+++ client/components/forms/index.ts
@@ -2,3 +2,4 @@ export * from './LoginForm';
 export * from './ProfileEditForm';
 export * from './FormField';
 export * from './ImagePicker';
+export * from './WorkoutForm';
