+++ client/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./AuthContext";
 export * from "./useFirebaseStorage";
+export * from "./useWorkout";
